requirements.txt
transformers==4.20.1
einops==0.4.1
tensorboard==2.9.1
pybind11==2.10.0
dm_control
functorch
autorom==0.4.2
ale-py==0.7
nltk==3.7
pycocoevalcap
pybind11
deepspeed==0.6.7
dm-tree
dm-env
traitlets
procgen==0.10.7
gym==0.25.0
gym_minigrid==1.0.3